Updated for Build 1.1.2296.24 + LazyCache work as follows. * Expanded weighting and started reworking Unit weights * Custom Implementation of ActorInfo, MonsterInfo * Fix for detecting and avoiding bad DiaUnits * Performance Improvements * Added methods for CacheBase .ToTrinityUnit() .ToTrinityObject() etc.

+++ Trinity/Items/CachedACDItem.cs
@@ -33,7 +33,7 @@ namespace Trinity
         public TrinityItemType TrinityItemType { get; set; }
         public FollowerType FollowerType { get; set; }
         public bool IsUnidentified { get; set; }
-        public int ItemStackQuantity { get; set; }
+        public long ItemStackQuantity { get; set; }
         public float Dexterity { get; set; }
         public float Intelligence { get; set; }
         public float Strength { get; set; }
